requirements.txt
mysql-connector-python~=8.0.33
fastapi~=0.95.1